requirements.txt
pandas
streamlit
openpyxl
google-generativeai
python-dotenv